requirements.txt
# Dependencias principales
requests>=2.28.0
python-dotenv>=0.19.0
pyyaml>=6.0

# Utilidades
click>=8.0.0
colorama>=0.4.4
tqdm>=4.64.0

# Base de datos
psycopg2-binary>=2.9.0

# Desarrollo y testing (opcional)
pytest>=7.0.0
pytest-cov>=4.0.0
black>=22.0.0
flake8>=5.0.0

# Análisis de datos (opcional)
pandas>=1.5.0
openpyxl>=3.0.10

# Notebooks (opcional)
jupyter>=1.0.0
ipython>=8.0.0
